+++ src/components/pages/Root/index.js
@@ -3,13 +3,13 @@ import PropTypes from 'prop-types';
 
 import Context from "@prisma-cms/context";
 import PrismaCmsComponent from "@prisma-cms/component";
-import FrontEditor from "../../../../App";
+import FrontEditor from "../../App";
 
 import { compose, graphql } from 'react-apollo';
 import gql from 'graphql-tag';
 import { Button } from 'material-ui';
 import PageHeader from './components/PageHeader';
-import RootConnector from '../../../../components/Root';
+import RootConnector from '../../Root';