+++ src/components/pages/Templates/index.js
@@ -10,7 +10,7 @@ import View from "./View";
 
 import Page from "../layout";
 
-// import FrontEditor from "../../../../App";
+// import FrontEditor from "../../App";
 
 class TemplatesPage extends Page {