+++ changelog.md
@@ -2,6 +2,7 @@
 =================================
 - Fix render children for void-elements
 - Fix templates page
+- Move pages into components
 
 1.11.3
 =================================